feat: move DeleteDialogTrigger component to etrian-registry for better organization
